requirements.txt
opencv-python
Pillow
colormath
torch
torchvision
torchaudio
matplotlib
PyYAML
requests
scipy
tqdm
pandas
seaborn
